requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
python-multipart
web3
qrcode
